+++ source/plugins/plgcommon/stkwlk.h
@@ -267,7 +267,7 @@ public:
   
   virtual void OnOSVersion(DWORD dwMajorVersion, DWORD dwMinorVersion, DWORD dwBuildNumber, LPCSTR szCSDVersion, WORD wServicePackMajor, WORD wServicePackMinor, WORD wSuiteMask, BYTE  wProductType)const;
   virtual void OnCallstackEntry(CallstackEntryType eType, CallstackEntry &entry, STACKFRAME64 &stackInfo)const;  
-  virtual void OnOutput(LPCSTR szText, int len)const;
+  virtual void OnOutput(LPCSTR szText, size_t len)const;
   void OnOutput1(LPCSTR szText)const { OnOutput(szText, strlen(szText)); }
 
   StackWalkerInternal *m_sw;
